requirements.txt
django==4.*
gunicorn
django-environ
djangorestframework
django-cors-headers
requests
psycopg2-binary

# TEMPLATE
django-ash

# Crypto
mnemonic
cryptography
environ
